feature/step-project step 1 fix code structure and some wrongs
