+++ src/main/java/az/iktlab/happyMiniProject/step1/Pet.java
@@ -1,4 +1,4 @@
-package az.iktlab.HappyMiniProject.step1;
+package az.iktlab.happyMiniProject.step1;
 
 import java.util.Arrays;
 
@@ -9,12 +9,29 @@ public class Pet {
     public byte trickLevel;
     public String[] habits;
 
+    public Pet() {
+    }
+
+    public Pet(String species, String nickname) {
+        this.species = species;
+        this.nickname = nickname;
+    }
+
+    public Pet(String species, String nickname, byte age, byte trickLevel, String[] habits) {
+        this.species = species;
+        this.nickname = nickname;
+        this.age = age;
+        this.trickLevel = trickLevel;
+        this.habits = habits;
+    }
+
+
     public void eat() {
         System.out.println("I am eating");
     }
 
     public void respond() {
-        System.out.println("Hello, owner. I am" + nickname + ". I miss you!");
+        System.out.printf("Hello, owner. I am + %s I miss you!",nickname);
     }
 
     public void foul() {
@@ -32,19 +49,8 @@ public class Pet {
                 '}';
     }
 
-    public Pet(String species, String nickname) {
-        this.species = species;
-        this.nickname = nickname;
-    }
 
-    public Pet(String species, String nickname, byte age, byte trickLevel, String[] habits) {
-        this.species = species;
-        this.nickname = nickname;
-        this.age = age;
-        this.trickLevel = trickLevel;
-        this.habits = habits;
-    }
 
-    public Pet() {
-    }
+
+
 }
